refactor: Removal of storageSync mechanism
