+++ projects/core/src/cart/services/multi-cart-state-persistence.service.ts
@@ -46,7 +46,7 @@ export class MultiCartStatePersistenceService implements OnDestroy {
     );
   }
 
-  protected onRead(state: { active: string }) {
+  protected onRead(state: { active: string } | undefined) {
     this.store.dispatch(new CartActions.ClearCartState());
     if (state) {
       this.store.dispatch(new CartActions.SetActiveCartId(state.active));
