+++ projects/core/src/state/config/state-config.ts
@@ -17,33 +17,6 @@ export enum StateTransferType {
 })
 export abstract class StateConfig {
   state?: {
-    /**
-     * @deprecated Since 3.0, to synchronize state from the browser storage, use StatePersistentService
-     */
-    storageSync?: {
-      /**
-       * A key name for the data stored in `localStorage`.
-       * Default is `DEFAULT_LOCAL_STORAGE_KEY`.
-       */
-      localStorageKeyName?: string;
-      /**
-       * A key name for the data stored in `sessionStorage`.
-       * Default is `DEFAULT_SESSION_STORAGE_KEY`.
-       */
-      sessionStorageKeyName?: string;
-      /**
-       * A set of state keys that should be synced with the specified browser's storage.
-       */
-      keys?: {
-        [key: string]: StorageSyncType;
-      };
-      /**
-       * A set of keys not to sync with the specified browser's storage.
-       */
-      excludeKeys?: {
-        [key: string]: StorageSyncType;
-      };
-    };
     ssrTransfer?: {
       keys?: {
         /**
