+++ projects/core/src/state/index.ts
@@ -1,4 +1,3 @@
-export * from './config/default-state-config';
 export * from './config/state-config';
 export * from './event/index';
 export * from './services/state-persistence.service';
