+++ projects/core/src/state/services/state-persistence.service.ts
@@ -2,12 +2,12 @@ import { Injectable } from '@angular/core';
 import { Observable, of, Subscription } from 'rxjs';
 import { map, tap, withLatestFrom } from 'rxjs/operators';
 import { StorageSyncType } from '../../state/config/state-config';
+import { WindowRef } from '../../window/window-ref';
 import {
   getStorage,
   persistToStorage,
   readFromStorage,
-} from '../../state/reducers/storage-sync.reducer';
-import { WindowRef } from '../../window/window-ref';
+} from '../utils/browser-storage';
 
 @Injectable({
   providedIn: 'root',
@@ -46,36 +46,38 @@ export class StatePersistenceService {
     state$: Observable<T>;
     context$?: Observable<string | Array<string>>;
     storageType?: StorageSyncType;
-    onRead?: (stateFromStorage: T) => void;
+    onRead?: (stateFromStorage: T | undefined) => void;
   }): Subscription {
     const storage = getStorage(storageType, this.winRef);
 
     const subscriptions = new Subscription();
 
     // Do not change order of subscription! Read should happen before write on context change.
-    subscriptions.add(
-      context$
-        .pipe(
-          map((context) => {
-            return readFromStorage(
-              storage,
-              this.generateKeyWithContext(context, key)
-            ) as T;
-          }),
-          tap((state) => onRead(state))
-        )
-        .subscribe()
-    );
+    if (storage) {
+      subscriptions.add(
+        context$
+          .pipe(
+            map((context) => {
+              return readFromStorage(
+                storage,
+                this.generateKeyWithContext(context, key)
+              ) as T | undefined;
+            }),
+            tap((state) => onRead(state))
+          )
+          .subscribe()
+      );
 
-    subscriptions.add(
-      state$.pipe(withLatestFrom(context$)).subscribe(([state, context]) => {
-        persistToStorage(
-          this.generateKeyWithContext(context, key),
-          state,
-          storage
-        );
-      })
-    );
+      subscriptions.add(
+        state$.pipe(withLatestFrom(context$)).subscribe(([state, context]) => {
+          persistToStorage(
+            this.generateKeyWithContext(context, key),
+            state,
+            storage
+          );
+        })
+      );
+    }
 
     return subscriptions;
   }
@@ -98,19 +100,25 @@ export class StatePersistenceService {
     key: string;
     context?: string | Array<string>;
     storageType?: StorageSyncType;
-  }): T {
+  }): T | undefined {
     const storage = getStorage(storageType, this.winRef);
+
+    if (!storage) {
+      return undefined;
+    }
 
     return readFromStorage(
       storage,
       this.generateKeyWithContext(context, key)
-    ) as T;
+    ) as T | undefined;
   }
 
   protected generateKeyWithContext(
     context: string | Array<string>,
     key: string
   ): string {
-    return `spartacus⚿${[].concat(context).join('⚿')}⚿${key}`;
+    return `spartacus⚿${([] as Array<string>)
+      .concat(context)
+      .join('⚿')}⚿${key}`;
   }
 }
